requirements.txt
# 🌐 Web Scraping & Automation
playwright==1.46.0
selenium==4.19.0
beautifulsoup4==4.12.3
lxml==5.2.1
requests==2.31.0
firecrawl-py==4.12.0
aiohttp==3.9.5
aiofiles==23.2.1

# 🤖 AI & ML
cohere==4.51
chromadb==0.4.24
langchain==0.1.20
langchain-community==0.0.34
openai==1.25.1
tiktoken==0.7.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
transformers==4.39.0
torch==2.2.1
tokenizers==0.19.1

# 🔍 Vector Search & Embeddings
hnswlib==0.7.0
annoy==1.17.3

# 📊 Data Processing
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.0
pydantic-settings==2.2.1

# 📁 File Processing
pickle5==0.0.12
pyarrow==15.0.2
orjson==3.10.0
msgpack==1.0.8

# 🚀 Web Framework & API
fastapi==0.110.0
uvicorn[standard]==0.29.0
flask==3.0.3
starlette==0.36.3
websockets==13.0
email-validator==2.1.1
slowapi==0.1.8
fastapi-limiter==0.1.6

# 🔧 Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
asyncio==3.4.3
tenacity==8.2.3
tqdm==4.66.2
joblib==1.3.2
scikit-learn==1.4.1.post1
watchfiles==0.21.0
python-dateutil==2.9.0.post0
pytz==2024.1

# 📦 Development
black==24.3.0
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-playwright==0.4.3
pytest-cov==4.1.0
ipython==8.23.0
jupyter==1.0.0
pytest-httpx==0.27.0
pytest-testdox==3.1.0
autopep8==2.0.4
mypy==1.9.0
pre-commit==3.7.0

# 🗄️ Database & Storage
sqlalchemy==2.0.29
alembic==1.13.1
duckdb==0.10.0
pysqlite3==0.5.3
redis==5.0.6
aioredis==2.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pymongo==4.6.3

# 📈 Visualization (opcjonalnie)
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# 🎯 CLI & Logging
click==8.1.7
rich==13.7.0
loguru==0.7.2
colorama==0.4.6
structlog==24.1.0
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.43b0

# 🔐 Security
cryptography==42.0.5
bcrypt==4.1.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
argon2-cffi==23.1.0
python-keycloak==2.13.1

# 🔄 Async & Concurrency
anyio==4.3.0
asgiref==3.8.1
uvloop==0.19.0

# 📡 Web & HTTP
httptools==0.6.1
itsdangerous==2.1.2
gunicorn==21.2.0
hypercorn==0.16.0

# 🧪 Testing & Mock
factory-boy==3.3.0
faker==24.9.0
responses==0.25.0
vcrpy==6.0.1
pytest-mock==3.14.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0

# 🛠️ DevOps & Deployment
dynaconf==3.2.4
celery==5.3.6
flower==2.0.1
sentry-sdk==1.40.6
opencensus==0.11.4
healthcheck==2.9.0

# 📚 Documentation
mkdocs==1.5.3
mkdocs-material==9.5.7
mkdocstrings[python]==0.24.0
typer==0.9.0

# 🌍 Internationalization
babel==2.14.0
polib==1.2.0